Refacto of 'crypto_helper' module
