+++ Makefile
@@ -1,4 +1,4 @@
-COMPONENTS = mithril-core mithril-aggregator mithril-client mithril-signer demo/protocol-demo mithril-test-lab
+COMPONENTS = mithril-common mithril-core mithril-aggregator mithril-client mithril-signer demo/protocol-demo mithril-test-lab
 GOALS := $(or $(MAKECMDGOALS),all)
 
 .PHONY: $(GOALS) $(COMPONENTS)
